implementing the triangulation, based on sweep-line/sweep-circle

+++ ced_cpu/include/Point.inl
@@ -1,6 +1,6 @@
 template <typename T>
 Point::Point(const T _x, const T _y)
 {
-    x = (int)std::move(_x);
-    y = (int)std::move(_y);
+    x = static_cast<float>(std::move(_x));
+    y = static_cast<float>(std::move(_y));
 }
